Fixed storm cell area calculation.
